working through TODO list items
