+++ index.html
@@ -31,14 +31,12 @@
                         <a class="nav-link" href="https://github.com/episphere/quest-dev" target="_blank">Source</a>
                     </li>
                     <li class="nav-item">
-                        <!-- TODO: -->
                         <a class="nav-link" href="https://github.com/episphere/quest/wiki" target="_blank">Wiki</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="https://github.com/episphere/quest-dev/issues" target="_blank">Issues</a>
                     </li>
                     <li class="nav-item">
-                        <!-- TODO: -->
                         <a class="nav-link" href="https://gitter.im/episphere/quest" target="_blank">Talk to us</a>
                     </li>
                     <li class="nav-item">
